face-emotion-model/requirements.txt
numpy==1.23.5
tensorflow==2.11.0
opencv-python==4.6.0.66
matplotlib==3.7.1
scikit-learn==1.2.2
pandas==1.5.3
tqdm==4.65.0
pillow==9.5.0
pyyaml==6.0
imbalanced-learn==0.11.0
